prepare for non-local envs

+++ serve.sh
@@ -1,14 +1,18 @@
 #!/bin/sh
 
-composer i && npm i
-rm -rf public/storage
-php artisan storage:link
-php artisan migrate
+if [ "$1" != "--only-serve" ]; then
+    composer i && npm i
+    rm -rf public/storage
+    php artisan storage:link
+    php artisan migrate
+fi
 
 stop_commands() {
   kill %1 %2
 }
 
-exec php artisan serve & exec npm run dev &
+folder=$(git rev-parse --show-toplevel)
+
+exec php "$folder"/artisan serve --host=0.0.0.0 & exec npm run dev &
 trap stop_commands INT
 wait
